requirements.txt
asgiref==3.2.2
asn1crypto==1.1.0
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
cffi==1.12.3
channels==2.3.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.6
django-environ==0.4.5
django-qrcode==0.3
django-user-agents==0.4.0
djangorestframework==3.10.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Pillow==6.2.0
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
ua-parser==0.8.0
user-agents==2.0
zope.interface==4.6.0
